Firebase Fix to "Firebase is not defined"
I got a good chance to investigate this problem this week. The problem is due to the Firebase upgrade to 3.x. There is documentation at https://firebase.google.com/support/guides/firebase-web. The point is that code of new Firebase is outdated, simply use firebase.database().ref(). I only have to change map.js to fix that.

The strange thing is that I did not get the same error as yours until I changed the code referenced to www.gstatic.com/firebasejs/3.5.1. So my first instinct is wrong, about not having the proper reference to that link.
I am using a different way to add/commit code changes. I used the web Github, again, and this time, I commit changes from the <>Code tab. I uploaded all 5 files for your convenience.

I was hoping to make this code fix earlier. Eager to see the change now :-)