Fix Warning on middle-container of the mainToolbar
Note: This small code fix is not related to the original problem, though it helps to fix this bit annoying bug.

Problem: 
In app.js, function window.addEventListener(), there is a warning on 
"Uncaught TypeError: Illegal invocation on…” from code Polymer.Base.transform('scale(' + scaleBottom + ') translateZ(0)', bottomContainer);
Since the code throws an error/warning, it does not execute the next code. My code fix will solve it.

The warning problem occurs when the main toolbar collapses/expands, affecting the middle and bottom container. On a desktop, this is done by moving/dragging the vertical scroll bar. On a phone, this is done by simply dragging the map with a finger. So this code is executed often.

Solution: 
The root cause of this problem is the code querySelector('#mainToolbar .bottom-container').
The paper-toolbar (mainToolbar in index.html) don't have a bottom container since it is not defined in that component. However the middl