Add files via upload
I made small minor code fixes. app.js (fix bottom-container issue) and map.js (reporting error from calling update()).

Section 1:
Before I mention the possible solution, let’s try to see the run-time error. The problem with the reserve/release feature is with the call addressRef.update() in map.js. As I said, I see this error on the existing app released from Github (https://easychargeapp.firebaseapp.com). 

The error that I only see was “Synchronous XMLHttpRequest on the main thread…”. However database updates are naturally asynchronous, not synchronous as the error says.  So if you get an error from the update() call, you would not see it immediately since it is asynchronous. And worse, with the Chrome devtool, the error is actually erased from the log when the web page is refreshed (after the update() call), hence I do not see the error; perhaps that’s what you experienced. You can see the error if you set the breakpoint on the right place. In the following, here are m